feat(calendar): calendar 일정 조회 dto 추가

+++ Back/canvEarth/src/main/java/com/b301/canvearth/domain/calendar/dto/CalendarResponseGetDto.java
@@ -4,20 +4,17 @@ import lombok.Builder;
 import lombok.Getter;
 import lombok.ToString;
 
-import java.util.Date;
-import java.util.List;
-
 @Getter
 @ToString
 public class CalendarResponseGetDto {
-    private final int calendarId;
+    private final long calendarId;
     private final String userId;
     private final String title;
     private final String startDate;
     private final String endDate;
 
     @Builder
-    public CalendarResponseGetDto(int calendarId, String userId, String title,
+    public CalendarResponseGetDto(long calendarId, String userId, String title,
                                   String startDate, String endDate) {
         this.calendarId = calendarId;
         this.userId = userId;
